requirement.txt
Django==2.0.7
django-cors-headers==2.4.0
django-filter==2.0.0
django-mathfilters==0.4.0
django-messages==0.5.3
djangorestframework==3.8.2
html5lib==1.0b8
Markdown==2.2
mysqlclient==1.3.13
Pillow==5.2.0
#pkg-resources==0.0.0
PyPDF2==1.26.0
pytz==2018.5
reportlab==3.5.8
six==1.11.0
xhtml2pdf==0.2.3
